rename.sh：文件已存在时提醒且不更名，防止覆盖已有表情。

+++ src/tpl/classic/img/face/rename.sh
@@ -5,6 +5,11 @@ do
 	baseName=${file%%.*}
 	hexName=`echo -n $baseName | xxd -ps`
 	newName=$hexName.gif
-	printf "%-15s ->     %s\n" $file $newName
-	mv $file $newName
+
+	if [ -e $newName ]; then
+		printf "%-15s -X     %s     (文件已存在)\n" $file $newName
+	else
+		printf "%-15s ->     %s\n" $file $newName
+		mv $file $newName
+	fi
 done
